authorization-app/requirements.txt
Flask==2.1.2
PyJWT==2.4.0
